+++ ethereum/db/base_db.py
@@ -1,0 +1,2 @@
+class BaseDB(object):
+    pass
