+++ ethereum/db/ephem_db.py
@@ -1,0 +1,53 @@
+from ethereum import utils
+from rlp.utils import str_to_bytes
+
+from .base_db import BaseDB
+
+class EphemDB(BaseDB):
+
+    def __init__(self):
+        self.db = {}
+        self.kv = self.db
+
+    def get(self, key):
+        return self.db[key]
+
+    def put(self, key, value):
+        self.db[key] = value
+
+    def delete(self, key):
+        del self.db[key]
+
+    def commit(self):
+        pass
+
+    def _has_key(self, key):
+        return key in self.db
+
+    def __contains__(self, key):
+        return self._has_key(key)
+
+    def __eq__(self, other):
+        return isinstance(other, self.__class__) and self.db == other.db
+
+    def __hash__(self):
+        return utils.big_endian_to_int(str_to_bytes(self.__repr__()))
+
+    def inc_refcount(self, key, value):
+        self.put(key, value)
+
+    def dec_refcount(self, key):
+        pass
+
+    def revert_refcount_changes(self, epoch):
+        pass
+
+    def commit_refcount_changes(self, epoch):
+        pass
+
+    def cleanup(self, epoch):
+        pass
+
+    def put_temporarily(self, key, value):
+        self.inc_refcount(key, value)
+        self.dec_refcount(key)
