+++ ethereum/db/listening_db.py
@@ -1,0 +1,34 @@
+from .base_db import BaseDB
+
+# Used for SPV proof creation
+class ListeningDB(BaseDB):
+
+    def __init__(self, db):
+        self.parent = db
+        self.kv = {}
+
+    def get(self, key):
+        if key not in self.kv:
+            self.kv[key] = self.parent.get(key)
+        return self.parent.get(key)
+
+    def put(self, key, value):
+        self.parent.put(key, value)
+
+    def commit(self):
+        pass
+
+    def delete(self, key):
+        self.parent.delete(key)
+
+    def _has_key(self, key):
+        return self.parent._has_key(key)
+
+    def __contains__(self, key):
+        return self.parent.__contains__(key)
+
+    def __eq__(self, other):
+        return self.parent == other
+
+    def __hash__(self):
+        return self.parent.__hash__()
